+++ arch/arm/mach-snapdragon/board.c
@@ -10,14 +10,10 @@
 #define pr_fmt(fmt) "QCOM: " fmt
 
 #include <asm/armv8/mmu.h>
-#include <asm/gpio.h>
 #include <asm/io.h>
 #include <asm/psci.h>
 #include <asm/system.h>
-#include <dm/device.h>
-#include <dm/pinctrl.h>
-#include <dm/uclass-internal.h>
-#include <dm/read.h>
+#include <dm/ofnode.h>
 #include <power/regulator.h>
 #include <env.h>
 #include <fdt_support.h>
@@ -29,7 +25,6 @@
 #include <lmb.h>
 #include <malloc.h>
 #include <fdt_support.h>
-#include <usb.h>
 #include <soc/qcom/smem.h>
 #include <sort.h>
 #include <soc/qcom/smem.h>
@@ -149,44 +144,6 @@ int board_fdt_blob_setup(void **fdtp)
 	qcom_psci_fixup(*fdtp);
 
 	return ret;
-}
-
-/*
- * Some Qualcomm boards require GPIO configuration when switching USB modes.
- * Support setting this configuration via pinctrl state.
- */
-int board_usb_init(int index, enum usb_init_type init)
-{
-	struct udevice *usb;
-	int ret = 0;
-
-	/* USB device */
-	ret = uclass_find_device_by_seq(UCLASS_USB, index, &usb);
-	if (ret) {
-		printf("Cannot find USB device\n");
-		return ret;
-	}
-
-	ret = dev_read_stringlist_search(usb, "pinctrl-names",
-					 "device");
-	/* No "device" pinctrl state, so just bail */
-	if (ret < 0)
-		return 0;
-
-	/* Select "default" or "device" pinctrl */
-	switch (init) {
-	case USB_INIT_HOST:
-		pinctrl_select_state(usb, "default");
-		break;
-	case USB_INIT_DEVICE:
-		pinctrl_select_state(usb, "device");
-		break;
-	default:
-		debug("Unknown usb_init_type %d\n", init);
-		break;
-	}
-
-	return 0;
 }
 
 /*
